g-setting

+++ database/migrations/2023_12_27_114518_create_general_settings_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('general_settings', function (Blueprint $table) {
             $table->id();
             $table->string('school_name');
-            $table->integer('school_phone');
+            $table->unsignedBigInteger('school_phone');
             $table->string('school_email');
             $table->string('school_logo_url');
             $table->string('school_address');
